fixed commit

+++ src/test/java/order/CreateOrdersTest.java
@@ -2,8 +2,9 @@ package order;
 
 import io.qameta.allure.junit4.DisplayName;
 import org.example.order.OrderClient;
-import org.example.order.OrderRequest;
+import org.junit.After;
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 import user.Credentials;
 import user.User;
@@ -20,44 +21,46 @@ public class CreateOrdersTest {
     private final UserGenerator generator = new UserGenerator();
     private User user;
     private UserClient client;
-    private OrderRequest orderRequest;
     String accessToken;
+
+    @Before
+    @DisplayName("Создание пользователя со случайным name")
+    public void setUp(){
+        client = new UserClient();
+    }
+    @After
+       public void cleanUp() {
+        if (accessToken != null) {
+            int statusCode = client.deleteUser(accessToken).extract().statusCode(); //удалить пользователя
+            Assert.assertEquals(SC_ACCEPTED, statusCode); //проверить, что пользователь успешно удалился
+        }
+    }
 
     @Test
     @DisplayName("Создание заказа авторизированным пользователем и инградиентами")
     public void createOrderUsersWithToken(){
-        client = new UserClient();
-        orderRequest = new OrderRequest();
         user = generator.random();
         client.createUser(user);
         var logInResponseLn = client.logIn(Credentials.from(user));
         accessToken = logInResponseLn.extract().path("accessToken");
         Assert.assertNotNull(accessToken);
-        var createOrderUsersWithTokenResponse = createOrderWithIngredients();
+        var createOrderUsersWithTokenResponse = createOrderWithIngredients(); // создать заказ
         var requestCreateOrderSuccess =  orderClient.createOrder(createOrderUsersWithTokenResponse, accessToken);
         int StatusCodeCreateOrderSuccess = requestCreateOrderSuccess.extract().statusCode();
         Assert.assertEquals(SC_OK, StatusCodeCreateOrderSuccess);
-        int statusCode  = client.deleteUser(accessToken).extract().statusCode(); //удалить пользователя
-        Assert.assertEquals(SC_ACCEPTED, statusCode); //проверить, что пользователь успешно удалился
-
     }
     @Test
     @DisplayName("Создание заказа неавторизированным пользователем с инградиентами")
     public void createOrderUsersWithoutToken(){
-        client = new UserClient();
-        orderRequest = new OrderRequest();
         var createOrderUsersWithOutTokenResponse = createOrderWithIngredients();
+        accessToken = null;
         var requestCreateOrderUsersWithOutToken = orderClient.createOrderWithoutTokenOrIngredients(createOrderUsersWithOutTokenResponse);
         int statusCodeRequestCreateOrderUsersWithOutToken = requestCreateOrderUsersWithOutToken.extract().statusCode();
-        System.out.println(statusCodeRequestCreateOrderUsersWithOutToken);
         Assert.assertEquals(SC_BAD_REQUEST, statusCodeRequestCreateOrderUsersWithOutToken);
     }
-
     @Test
     @DisplayName("Создание заказа авторизированным пользователем, но без инградиентов")
     public void createOrderUsersWithoutIngredients(){
-        client = new UserClient();
-        orderRequest = new OrderRequest();
         user = generator.random();
         client.createUser(user);
         var logInResponseLn = client.logIn(Credentials.from(user));
@@ -67,15 +70,10 @@ public class CreateOrdersTest {
         var responseCreateOrderUsersWithIn = orderClient.createOrderWithoutTokenOrIngredients(createOrderUsersWithIn);
         int statusCode = responseCreateOrderUsersWithIn.extract().statusCode();
         Assert.assertEquals(SC_BAD_REQUEST, statusCode);
-        client.deleteUser(accessToken).extract().statusCode(); //удалить пользователя
-
     }
-
     @Test
     @DisplayName("Создание заказа авторизированным пользователем, но с инградиентами, которых нет")
     public void createOrderUsersWithWrongIngredients(){
-        client = new UserClient();
-        orderRequest = new OrderRequest();
         user = generator.random();
         client.createUser(user);
         var logInResponseLn = client.logIn(Credentials.from(user));
@@ -85,7 +83,6 @@ public class CreateOrdersTest {
         var requestCreateWithWrongIngredients = orderClient.createOrderWithoutTokenOrIngredients(createOrderUsersWrongIngredients);
         int statusCode = requestCreateWithWrongIngredients.extract().statusCode();
         Assert.assertEquals(SC_BAD_REQUEST, statusCode);
-        client.deleteUser(accessToken).extract().statusCode(); //удалить пользователя
     }
 
 }
