+++ src/test/java/user/CreateUserTest.java
@@ -2,16 +2,25 @@ package user;
 
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import com.github.javafaker.Faker;
 
 import static org.apache.http.HttpStatus.*;
 
-public class CreateUserTest {
+public class CreateUserTest extends UserClient{
     private final UserGenerator generator = new UserGenerator();
+    private ValidatableResponse createResponse;
+
+
     private User user = new User();
+    Faker faker = new Faker();
+    String name = faker.funnyName().name();
+    String password = faker.internet().password();
     private UserClient client;
+    String accessToken;
 
     @Before
     @DisplayName("Создание пользователя со случайным name")
@@ -19,20 +28,22 @@ public class CreateUserTest {
         client = new UserClient();
         user = generator.random();
     }
-
-    @Test
-    @DisplayName("Успешная регистрация пользователя")
-    public void registrationUserTest(){
-        var createResponseCr = client.createUser(user);
-        int statusCodeCr = createResponseCr.extract().statusCode();
-        String accessToken = createResponseCr.extract().path("accessToken");
+     @After
+     public void cleanUp() {
+        if (accessToken != null) {
+            int statusCode = client.deleteUser(accessToken).extract().statusCode(); //удалить пользователя
+            Assert.assertEquals(SC_ACCEPTED, statusCode); //проверить, что пользователь успешно удалился
+        }
+     }
+     @Test
+     @DisplayName("Успешная регистрация пользователя")
+     public void registrationUserTest(){
+        var createResponse = client.createUser(user);
+        int statusCodeCr = createResponse.extract().statusCode();
+        accessToken = createResponse.extract().path("accessToken");
         Assert.assertNotNull(accessToken);
         Assert.assertEquals(SC_OK, statusCodeCr);
-
-        int statusCode  = client.deleteUser(accessToken).extract().statusCode(); //удалить пользователя
-        Assert.assertEquals(SC_ACCEPTED, statusCode); //проверить, что пользователь успешно удалился
-
-    }
+     }
     @Test
     @DisplayName("Создание пользователя с существующим логином")
     public void canNotCreateTwoSameUser(){
@@ -41,30 +52,24 @@ public class CreateUserTest {
         int statusCodeSameCourier = courierResponse.extract().statusCode(); //получение кода ответа
         Assert.assertEquals(SC_FORBIDDEN, statusCodeSameCourier);
         if (statusCodeSameCourier == 201) {
-            String accessToken = courierResponse.extract().path("accessToken");
-            Assert.assertNotNull(accessToken);
-            Assert.assertEquals(SC_OK, statusCodeSameCourier);
-            int statusCode  = client.deleteUser(accessToken).extract().statusCode(); //удалить пользователя
-            Assert.assertEquals(SC_ACCEPTED, statusCode); //проверить, что пользователь успешно удалился
-
-        }
+        accessToken = courierResponse.extract().path("accessToken");
+        Assert.assertNotNull(accessToken);
+        Assert.assertEquals(SC_OK, statusCodeSameCourier);
+         }
     }
     @Test
     @DisplayName("Создание пользователя с пустым логином")
     public void canNotCreateCourierNoLogin() {
-        User user = new User("", "T&^&^*9987!", "Test22");
+        User user = new User("", password, name);
         ValidatableResponse courierResponse = client.createUser(user);
         int statusCodeNoLogin = courierResponse.extract().statusCode(); //получение кода ответа
         Assert.assertEquals(SC_FORBIDDEN, statusCodeNoLogin);
         if (statusCodeNoLogin == 201) {
-            String accessToken = courierResponse.extract().path("accessToken");
+            accessToken = courierResponse.extract().path("accessToken");
             Assert.assertNotNull(accessToken);
             Assert.assertEquals(SC_OK, statusCodeNoLogin);
-            int statusCode  = client.deleteUser(accessToken).extract().statusCode(); //удалить пользователя
-            Assert.assertEquals(SC_ACCEPTED, statusCode); //проверить, что пользователь успешно удалился
         }
     }
-
     @Test
     @DisplayName("Создание пользователя без пароля")
     public void canNotCreateCourierWithoutPassword() {
@@ -74,11 +79,9 @@ public class CreateUserTest {
         System.out.println(statusCodeWithoutPass);
         Assert.assertEquals(SC_FORBIDDEN, statusCodeWithoutPass);
         if (statusCodeWithoutPass == 201) {
-            String accessToken = courierResponse.extract().path("accessToken");
+            accessToken = courierResponse.extract().path("accessToken");
             Assert.assertNotNull(accessToken);
             Assert.assertEquals(SC_OK, statusCodeWithoutPass);
-            int statusCode  = client.deleteUser(accessToken).extract().statusCode(); //удалить пользователя
-            Assert.assertEquals(SC_ACCEPTED, statusCode); //проверить, что пользователь успешно удалился
         }
     }
 }
